Aggiunto controllo per PS1=""

+++ setup.sh
@@ -41,6 +41,7 @@ do
   yes_no || break;
 
 done
+[[ $NEW_PS1 == "" ]] && { echo -e "Attenzione il prompt è vuoto!\nVuoi proseguire?"; yes_no|| exit 1; }
 
 echo "export PS1=\"$NEW_PS1\" " > ~/.prompt
 if ! grep -q "[ -f ~/.prompt ] &&  . ~/.prompt" ~/.bash_profile
